+++ src/main/java/com/app/user/entity/VehicleInspectionDetailsEntity.java
@@ -47,7 +47,7 @@ public class VehicleInspectionDetailsEntity {
 	private int leastTireThicknessAllowed;
 
 	@Column(name = "odometer_reading_when_fitted", nullable = true)
-	private Integer odometerReadingWhenFitted;
+	private Double odometerReadingWhenFitted;
 
 	@Column(name = "tire_pattern", nullable = false)
 	private String tirePattern;
@@ -56,25 +56,25 @@ public class VehicleInspectionDetailsEntity {
 	private Date tireRemovalDate;
 
 	@Column(name = "odometer_reading_when_removed", nullable = true)
-	private Integer odometerReadingWhenRemoved;
+	private Double odometerReadingWhenRemoved;
 
 	@Column(name = "reco_pressure", nullable = false)
-	private int recoPressure;
+	private Double recoPressure;
 
 	@Column(name = "obs_pressure", nullable = false)
-	private int obsPressure;
+	private Double obsPressure;
 
 	@Column(name = "otd_mm", nullable = false)
-	private Integer otdMm;
+	private Double otdMm;
 
 	@Column(name = "lst_mm", nullable = false)
-	private Integer lstMm;
+	private Double lstMm;
 
 	@Column(name = "ct_mm", nullable = false)
-	private Integer ctMm;
+	private Double ctMm;
 
 	@Column(name = "rst_mm", nullable = false)
-	private Integer rstMm;
+	private Double rstMm;
 
 	@Column(name = "observation_category", nullable = false)
 	private Long observationCategory;
@@ -107,7 +107,7 @@ public class VehicleInspectionDetailsEntity {
 	private String wearAnalysis;
 
 	@Column(name = "rtd", nullable = true)
-	private Integer rtd;
+	private Double rtd;
 
 	@Column(name = "projected_mileage", nullable = true)
 	private Double projectedMileage;
@@ -203,11 +203,11 @@ public class VehicleInspectionDetailsEntity {
 		this.leastTireThicknessAllowed = leastTireThicknessAllowed;
 	}
 
-	public Integer getOdometerReadingWhenFitted() {
+	public Double getOdometerReadingWhenFitted() {
 		return odometerReadingWhenFitted;
 	}
 
-	public void setOdometerReadingWhenFitted(Integer odometerReadingWhenFitted) {
+	public void setOdometerReadingWhenFitted(Double odometerReadingWhenFitted) {
 		this.odometerReadingWhenFitted = odometerReadingWhenFitted;
 	}
 
@@ -227,59 +227,59 @@ public class VehicleInspectionDetailsEntity {
 		this.tireRemovalDate = tireRemovalDate;
 	}
 
-	public Integer getOdometerReadingWhenRemoved() {
+	public Double getOdometerReadingWhenRemoved() {
 		return odometerReadingWhenRemoved;
 	}
 
-	public void setOdometerReadingWhenRemoved(Integer odometerReadingWhenRemoved) {
+	public void setOdometerReadingWhenRemoved(Double odometerReadingWhenRemoved) {
 		this.odometerReadingWhenRemoved = odometerReadingWhenRemoved;
 	}
 
-	public int getRecoPressure() {
+	public Double getRecoPressure() {
 		return recoPressure;
 	}
 
-	public void setRecoPressure(int recoPressure) {
+	public void setRecoPressure(Double recoPressure) {
 		this.recoPressure = recoPressure;
 	}
 
-	public int getObsPressure() {
+	public Double getObsPressure() {
 		return obsPressure;
 	}
 
-	public void setObsPressure(int obsPressure) {
+	public void setObsPressure(Double obsPressure) {
 		this.obsPressure = obsPressure;
 	}
 
-	public Integer getOtdMm() {
+	public Double getOtdMm() {
 		return otdMm;
 	}
 
-	public void setOtdMm(Integer otdMm) {
+	public void setOtdMm(Double otdMm) {
 		this.otdMm = otdMm;
 	}
 
-	public Integer getLstMm() {
+	public Double getLstMm() {
 		return lstMm;
 	}
 
-	public void setLstMm(Integer lstMm) {
+	public void setLstMm(Double lstMm) {
 		this.lstMm = lstMm;
 	}
 
-	public Integer getCtMm() {
+	public Double getCtMm() {
 		return ctMm;
 	}
 
-	public void setCtMm(Integer ctMm) {
+	public void setCtMm(Double ctMm) {
 		this.ctMm = ctMm;
 	}
 
-	public Integer getRstMm() {
+	public Double getRstMm() {
 		return rstMm;
 	}
 
-	public void setRstMm(Integer rstMm) {
+	public void setRstMm(Double rstMm) {
 		this.rstMm = rstMm;
 	}
 
@@ -363,11 +363,11 @@ public class VehicleInspectionDetailsEntity {
 		this.wearAnalysis = wearAnalysis;
 	}
 
-	public Integer getRtd() {
+	public Double getRtd() {
 		return rtd;
 	}
 
-	public void setRtd(Integer rtd) {
+	public void setRtd(Double rtd) {
 		this.rtd = rtd;
 	}
